+++ api/src/identity-access-management/domain/constants/oidc-errors.js
@@ -1,0 +1,6 @@
+export const OIDC_ERRORS = {
+  USER_INFO: {
+    missingFields: { shortCode: 'OIDC01', code: 'USER_INFO_MISSING_FIELDS' },
+    badResponseFormat: { shortCode: 'OIDC02', code: 'USER_INFO_BAD_RESPONSE_FORMAT' },
+  },
+};
