+++ api/src/identity-access-management/domain/services/oidc-authentication-service.js
@@ -6,12 +6,12 @@ import ms from 'ms';
 import * as client from 'openid-client';
 
 import { config } from '../../../shared/config.js';
-import { OIDC_ERRORS } from '../../../shared/constants.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { OidcError, OidcMissingFieldsError } from '../../../shared/domain/errors.js';
 import { AuthenticationSessionContent } from '../../../shared/domain/models/AuthenticationSessionContent.js';
 import { temporaryStorage } from '../../../shared/infrastructure/key-value-storages/index.js';
 import { logger } from '../../../shared/infrastructure/utils/logger.js';
+import { OIDC_ERRORS } from '../constants/oidc-errors.js';
 import { AuthenticationMethod } from '../models/AuthenticationMethod.js';
 import { ClaimManager } from '../models/ClaimManager.js';
