+++ api/tests/identity-access-management/unit/domain/services/oidc-authentication-service.test.js
@@ -2,11 +2,11 @@ import jsonwebtoken from 'jsonwebtoken';
 import ms from 'ms';
 import sinon from 'sinon';
 
+import { OIDC_ERRORS } from '../../../../../src/identity-access-management/domain/constants/oidc-errors.js';
 import { AuthenticationMethod } from '../../../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
 import { UserToCreate } from '../../../../../src/identity-access-management/domain/models/UserToCreate.js';
 import { OidcAuthenticationService } from '../../../../../src/identity-access-management/domain/services/oidc-authentication-service.js';
 import { config as settings } from '../../../../../src/shared/config.js';
-import { OIDC_ERRORS } from '../../../../../src/shared/constants.js';
 import { DomainTransaction } from '../../../../../src/shared/domain/DomainTransaction.js';
 import { OidcError, OidcMissingFieldsError } from '../../../../../src/shared/domain/errors.js';
 import { AuthenticationSessionContent } from '../../../../../src/shared/domain/models/AuthenticationSessionContent.js';
